refactor: rename mcp_outro.py to cogni_memory_mcp_outro.py
- More descriptive filename for the Cogni memory MCP outro helper
- Update import references in dolt_ops.py and autogen_work_reader.py
- Maintains same functionality with clearer naming
